summary.js extracted
